Fix not being able to enumerate objects in migration after deleting
Fixes https://github.com/realm/realm-cocoa/issues/3390

The cause of the issue is that when deleting objects, the number of objects does not match between the new schema and the old schema.

In the current version, this case crashes, so it will not be a big problem.
However, If inserting the same number of objects as deleted objects, it will not crash and can not be enumerated correctly (the order will not match between the old schema and the new schema), the issue becomes serious.

To solve this problem, defer deletion during migration and actually delete the object when migration finished.